feat: fetch desktop info
